Allow feature names that contain a dot
Previously the logic for parsing a column definition would fail if the definition contains more than one dot, which made it impossible to pull values from (document or annotation) features whose name contains a dot.  Now we limit the split to 2 parts, so everything after the first dot is the feature name even if that contains other dots.  In particular, this change means it is now possible to specify a column definition ".gate.SourceURL" to get the document URL as a column in the CSV.